+++ src/main/java/org/locationtech/proj4j/BasicCoordinateTransform.java
@@ -79,28 +79,45 @@ public class BasicCoordinateTransform implements CoordinateTransform {
         doDatumTransform = doInverseProjection && doForwardProjection
                 && srcCRS.getDatum() != tgtCRS.getDatum();
 
+        boolean geocentric = false;
+
         if (doDatumTransform) {
 
             boolean isEllipsoidEqual = srcCRS.getDatum().getEllipsoid().isEqual(tgtCRS.getDatum().getEllipsoid());
-            transformViaGeocentric = ! isEllipsoidEqual || srcCRS.getDatum().hasTransformToWGS84()
+            geocentric = ! isEllipsoidEqual || srcCRS.getDatum().hasTransformToWGS84()
                     || tgtCRS.getDatum().hasTransformToWGS84();
 
-            if (transformViaGeocentric) {
+            if (geocentric) {
                 srcGeoConv = new GeocentricConverter(srcCRS.getDatum().getEllipsoid());
                 tgtGeoConv = new GeocentricConverter(tgtCRS.getDatum().getEllipsoid());
 
-                if (srcCRS.getDatum().getTransformType() == Datum.TYPE_GRIDSHIFT) {
-                    srcGeoConv.overrideWithWGS84Params();
-                }
+                int srcTransformType = srcCRS.getDatum().getTransformType();
+                int tgtTransformType = tgtCRS.getDatum().getTransformType();
 
-                if (tgtCRS.getDatum().getTransformType() == Datum.TYPE_GRIDSHIFT) {
-                    tgtGeoConv.overrideWithWGS84Params();
+                if (srcTransformType == Datum.TYPE_GRIDSHIFT || tgtTransformType == Datum.TYPE_GRIDSHIFT) {
+
+	                if (srcTransformType == Datum.TYPE_GRIDSHIFT) {
+	                    srcGeoConv.overrideWithWGS84Params();
+	                }
+
+	                if (tgtTransformType == Datum.TYPE_GRIDSHIFT) {
+	                    tgtGeoConv.overrideWithWGS84Params();
+	                }
+
+	                // After WGS84 params override, check if geocentric transform is still required
+	                // https://github.com/OSGeo/PROJ/blob/5.2.0/src/pj_transform.c#L892
+	                if(srcGeoConv.isEqual(tgtGeoConv)) {
+	                    geocentric = false;
+	                    srcGeoConv = null;
+	                    tgtGeoConv = null;
+	                }
+
                 }
             }
 
-        } else {
-        	transformViaGeocentric=false;
         }
+
+        transformViaGeocentric = geocentric;
     }
 
     @Override
@@ -172,6 +189,10 @@ public class BasicCoordinateTransform implements CoordinateTransform {
                 || tgtCRS.getDatum().getTransformType() == Datum.TYPE_UNKNOWN)
             return;
 
+        /* -------------------------------------------------------------------- */
+        /*	If this datum requires grid shifts, then apply it to geodetic    */
+        /*      coordinates.                                                    */
+        /* -------------------------------------------------------------------- */
         if (srcCRS.getDatum().getTransformType() == Datum.TYPE_GRIDSHIFT) {
             srcCRS.getDatum().shift(pt);
         }
@@ -202,7 +223,9 @@ public class BasicCoordinateTransform implements CoordinateTransform {
             tgtGeoConv.convertGeocentricToGeodetic(pt);
         }
 
-
+        /* -------------------------------------------------------------------- */
+        /*      Apply grid shift to destination if required.                    */
+        /* -------------------------------------------------------------------- */
         if (tgtCRS.getDatum().getTransformType() == Datum.TYPE_GRIDSHIFT) {
             tgtCRS.getDatum().inverseShift(pt);
         }